extract group_context_extensions module to own file
